done with core java assignment

+++ core-java-assignment/src/com/revature/cja/SimpleMethods.java
@@ -1,8 +1,13 @@
+//Q15. Write a program that defines an interface having the following methods: addition,
+//subtraction, multiplication, and division. Create a class that implements this interface
+//and provides appropriate functionality to carry out the required operations. Hard code
+//two operands in a test class having a main method that calls the implementing class.
+
 package com.revature.cja;
 
 public interface SimpleMethods {
-	int addition(int num1, int num2);
-	int subtraction(int num1, int num2);
-	int multiplication(int num1, int num2);
-	int division(int num1, int num2);
+	double addition(double num1, double num2);
+	double subtraction(double num1, double num2);
+	double multiplication(double num1, double num2);
+	double division(double num1, double num2);
 }
